Loss/Perf/Sparsity data
Summary:
Loads data for a "test" project whenever creating a new project. A "test" folder is required in the projects folder on the server. This is until we implement loading the list of projects from the server.
Adds infrastructure to load loss and perf data for a project.
Adds redux sagas for side effects from selecting a new project.
Adds reselect for redux selectors.
Adds link from redux layer to the ui layer for sparse, dense and sparsity data.

@mark.kurtz The function to compute sparsity based on loss and perf is in ui/store/selectors/layers, called **sparsity**

Reviewers: mark.kurtz, kevinaer, alan

Reviewed By: mark.kurtz

Subscribers: mark.kurtz, PHID-PROJ-4wvim5epgiak4ulpeupd

Differential Revision: https://phabricator.neuralmagic.com:8021/D2098